Add "any" type for when the actual type can be more than one type
Sometimes, a field can be either a string or an integer or something
else. Let's add an escape hatch "any" type for being able to represent
these cases in IDL.

In the future we may want to add union types but for now an "any" type
will do the trick.